Conflicts:
	Platform/DirectX11/SimulCloudRendererDX1x.cpp
	Platform/PS4
	PlugIns/Vision/Editor/vForge/EditorPlugins/ThirdParty/TrueSkyPlugin/TrueSkyEditorPlugin/TrueSkyEditorPlugin_VC9.csproj
	PlugIns/Vision/Editor/vForge/EditorPlugins/ThirdParty/TrueSkyPlugin/TrueSkyManaged/WeatherInstance.h
	Products/TrueSky/Help/Help.rtf
	Simul_Trunk_VC9.sln

+++ Platform/DirectX11/SaveTextureDX1x.cpp
@@ -26,7 +26,7 @@ namespace simul
 				char number_text[10];
 				sprintf_s(number_text,10,"%d",number);
 				std::string nstr(number_text);
-				int pos=fn_utf8.find_last_of(".");
+				int pos=(int)fn_utf8.find_last_of(".");
 				fn_utf8=fn_utf8.replace(fn_utf8.begin()+pos,fn_utf8.begin()+pos+1,nstr+".");
 			}
 			std::wstring wfilename=simul::base::Utf8ToWString(fn_utf8);
